+++ src/plone/restapi/deserializer/local_roles.py
@@ -3,6 +3,7 @@ from AccessControl.interfaces import IRoleManager
 from plone.restapi.deserializer import json_body
 from plone.restapi.interfaces import IDeserializeFromJson
 from Products.CMFCore.interfaces import ICatalogAware
+from Products.CMFPlone.interfaces import IPloneSiteRoot
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.event import notify
@@ -54,7 +55,10 @@ class DeserializeFromJson(object):
                 user["roles"] = roles_list
             roles_reindex = sharing_view.update_role_settings(new_roles, reindex=False)
 
-        if ICatalogAware(self.context) and (inherit_reindex or roles_reindex):
+        # reindex object security
+        can_reindex = (ICatalogAware(self.context, None) or
+                       IPloneSiteRoot.providedBy(self.context))
+        if can_reindex and (inherit_reindex or roles_reindex):
             self.context.reindexObjectSecurity()
             if LOCALROLES_MODIFIED_EVENT_AVAILABLE:
                 notify(LocalrolesModifiedEvent(self.context, self.request))