+++ src/plone/restapi/tests/test_dxcontent_deserializer.py
@@ -91,6 +91,18 @@ class TestDXContentDeserializer(unittest.TestCase, OrderingMixin):
             self.event.descriptions[0].attributes,
         )
 
+    def test_deserializer_notifies_when_field_is_set_to_null(self):
+        def handler(obj, event):
+            obj._handler_called = True
+            self.event = event
+        provideHandler(handler, (IDexterityItem, IObjectModifiedEvent,))
+        self.deserialize(body='{"test_textline_field": null}')
+        self.assertTrue(getattr(self.portal.doc1, '_handler_called', False),
+                        'IObjectModifiedEvent not notified')
+        self.assertEqual(
+            ('IDXTestDocumentSchema.test_textline_field',),
+            self.event.descriptions[0].attributes)
+
     def test_deserializer_does_not_update_field_without_write_permission(self):
         self.portal.doc1.test_write_permission_field = u"Test Write Permission"
         setRoles(self.portal, TEST_USER_ID, ["Member", "Contributor", "Editor"])