+++ src/plone/restapi/types/utils.py
@@ -199,10 +199,22 @@ def get_jsonschema_for_portal_type(portal_type, context, request, excluded_field
     )
 
 
-def get_vocabulary_url(vocab_name, context, request):
+def get_vocab_like_url(endpoint, locator, context, request):
     try:
         context_url = context.absolute_url()
     except AttributeError:
         portal = getSite()
         context_url = portal.absolute_url()
-    return "{}/@vocabularies/{}".format(context_url, vocab_name)
+    return '/'.join((context_url, endpoint, locator))
+
+
+def get_vocabulary_url(vocab_name, context, request):
+    return get_vocab_like_url('@vocabularies', vocab_name, context, request)
+
+
+def get_querysource_url(field, context, request):
+    return get_vocab_like_url('@querysources', field.getName(), context, request)
+
+
+def get_source_url(field, context, request):
+    return get_vocab_like_url('@sources', field.getName(), context, request)